move reshape to decomposition
