+++ .github/settings/tests.json
@@ -32,7 +32,7 @@
     { "runs-on": "n150",   "image": "tracy",  "script": "ttnn_runtime.sh", "args": "runtime/test/ttnn/python/distributed/n150" },
     { "runs-on": "n300",   "image": "tracy",  "script": "ttnn_runtime.sh", "args": "runtime/test/ttnn/python/n300" },
     { "runs-on": "llmbox", "image": "tracy",  "script": "ttnn_runtime.sh", "args": "runtime/test/ttnn/python/distributed/llmbox" },
-    { "runs-on": "llmbox", "image": "speedy", "script": "ttrt.sh", "args": ["run", "Dialect/TTNN/trace/scenarios/llmbox", "--enable-program-cache", "--loops", "10"] },
+    { "runs-on": "llmbox", "image": "speedy", "script": "ttrt.sh", "args": ["run", "Dialect/TTNN/trace/scenarios/llmbox", "--enable-program-cache", "--loops", "3"] },
     { "runs-on": "n300",   "image": "tracy",  "script": "runtime_debug.sh" },
     { "runs-on": "n150",   "image": "speedy",  "script": "ttnn_standalone.sh" },
     { "runs-on": "n150",   "image": "tracy",  "script": "pykernel.sh" },
