+++ lib/Dialect/TTNN/Transforms/CMakeLists.txt
@@ -63,7 +63,6 @@ add_mlir_dialect_library(MLIRTTNNTransforms
         Workarounds/Decomposition/UpsampleOpRewritePattern.cpp
         Workarounds/Decomposition/RMSNormConfigRewritePattern.cpp
         Workarounds/Decomposition/TopKRouterGptDecompositionRewritePattern.cpp
-        Workarounds/Decomposition/DistributedRMSNormReshapeToCanonicalShapeRewritePattern.cpp
         Workarounds/Decomposition/DistributedRMSNormWidthShardInputRewritePattern.cpp
         Workarounds/Decomposition/LayerNormPreAllGatherZeroPadRewritePattern.cpp
         Workarounds/Decomposition/LayerNormPostAllGatherDecompositionRewritePattern.cpp
