+++ test/ttmlir/Dialect/TTNN/trace/scenarios/llmbox/llama_3_1_70b_tp_decode_layer.mlir
@@ -2,6 +2,7 @@
 // RUN: ttmlir-opt --ttir-to-ttnn-backend-pipeline="system-desc-path=%system_desc_path% mesh-shape=2,4 optimization-level=1 enable-trace=true experimental-weight-dtype=bfp_bf8 enable-const-eval=true enable-cpu-hoisted-const-eval=true" -o llama_3_1_70b_tp_decode_layer_ttnn.mlir %models/single_blocks_and_layers/llama_3_1_70b_tp_decode_layer.mlir
 // RUN: FileCheck %s --input-file=llama_3_1_70b_tp_decode_layer_ttnn.mlir
 // RUN: ttmlir-translate --ttnn-to-flatbuffer -o %t.ttnn llama_3_1_70b_tp_decode_layer_ttnn.mlir
+// RUN: ttrt run --enable-program-cache --loops 2 --trace-region-size 209715200 %t.ttnn
 // Verify the expected fused / distributed ops are emitted by the pipeline.
 
 // Input layernorm before the attention block.
